requirements.txt
aiohttp==3.7.1
amqp==5.0.1
appnope==0.1.0
argon2-cffi==20.1.0
async-timeout==3.0.1
attrs==20.1.0
aws==0.2.5
awscli==1.19.1
backcall==0.2.0
bcrypt==3.2.0
billiard==3.6.3.0
bleach==3.1.5
boto==2.49.0
boto3==1.17.1
botocore==1.20.1
cachetools==4.1.1
celery==5.0.1
certifi==2020.6.20
cffi==1.14.2
chain==1.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.3
cryptography==3.1.1
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
executing==0.5.2
fabric==2.5.0
fuzzywuzzy==0.18.0
google-api-core==1.22.0
google-api-python-client==1.12.8
google-auth==1.25.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gspread==3.6.0
httplib2==0.18.1
idna==2.10
invoke==1.4.1
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.1.0
jupyter-core==4.6.3
kombu==5.0.2
MarkupSafe==1.1.1
mistune==0.8.4
multidict==5.0.0
mysql-connector-python==8.0.21
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
numpy==1.19.0
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.4.0.44
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
paramiko==2.7.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prettytable==1.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.6.1
Pygments==2.6.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.6
QtPy==1.9.0
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.5
s3transfer==0.3.3
Send2Trash==1.5.0
sentry-sdk==0.17.4
six==1.15.0
SQLAlchemy==1.3.18
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.3
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.9
varname==0.4.0
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
yarl==1.6.2
zope.interface==5.1.0